% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data_berlin}
\alias{download_data_berlin}
\title{Download data to draw Berlin template maps}
\usage{
download_data_berlin(osmdata = TRUE)
}
\arguments{
\item{osmdata}{A Boolean. Should data from OpenStreetMap be downloaded?}
}
\value{
Geodata of Berlin.
}
\description{
Download data to draw Berlin template maps
}
\examples{
\dontrun{
download_data_berlin()
}

}
