% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fisbroker_atom.R
\name{download_fisbroker_atom}
\alias{download_fisbroker_atom}
\title{Download ATOM data from fisbroker database https://fbinter.stadt-berlin.de/fb/index.jsp
select a layer and click on the ATOM link. A pop-up window will open and you can copy the zip link and paste
it to this function.}
\usage{
download_fisbroker_atom(zip_link, path, name)
}
\arguments{
\item{zip_link}{the atom url of the data from fisbroker.}

\item{path}{path where the file is stored}

\item{name}{name of the folder where the data should be stored (will be created if it does not exists)}
}
\value{
A folder with the data as a geopackage.
}
\description{
Download ATOM data from fisbroker database https://fbinter.stadt-berlin.de/fb/index.jsp
select a layer and click on the ATOM link. A pop-up window will open and you can copy the zip link and paste
it to this function.
}
\examples{
\dontrun{
url <- "https://fbinter.stadt-berlin.de/fb/atom/Vegetationshoehen/veghoehe_2020.zip"
download_fisbroker_atom(zip_link = url, path = tempdir(), name = "vheights")
}
}
