% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fisbroker_wfs.R
\name{download_fisbroker_wfs}
\alias{download_fisbroker_wfs}
\title{Download WFS data from fisbroker database https://fbinter.stadt-berlin.de/fb/index.jsp
select a layer and click on the WFS link. A pop-up window will open and you can copy the link and paste
it to this function.}
\usage{
download_fisbroker_wfs(link)
}
\arguments{
\item{link}{the url of the data from fisbroker.}
}
\value{
A folder with the data as a geopackage.
}
\description{
Download WFS data from fisbroker database https://fbinter.stadt-berlin.de/fb/index.jsp
select a layer and click on the WFS link. A pop-up window will open and you can copy the link and paste
it to this function.
}
\examples{
\dontrun{
url <- "https://fbinter.stadt-berlin.de/fb/wfs/data/senstadt/s_wfs_alkis_bezirk"
d6berlin::download_fisbroker_wfs(link = url)
}
}
