% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_berlin}
\alias{sf_berlin}
\title{Berlin districts (WGS 84)}
\format{
An object of class \code{sfc_MULTIPOLYGON} (inherits from \code{sfc}) of length 1.
}
\source{
\url{https://daten.odis-berlin.de/en/dataset/bezirksgrenzen}
}
\usage{
sf_berlin
}
\description{
A shape file containing the shape of Berlin.
}
\keyword{datasets}
