% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_districts}
\alias{sf_districts}
\title{Berlin districts (WGS 84)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.
}
\source{
Data downloaded from \url{https://daten.odis-berlin.de/en/dataset/bezirksgrenzen}.
}
\usage{
sf_districts
}
\description{
An sf object containing the12 districts of Berlin as multipolygons. The variables are as follows:
}
\details{
\itemize{
\item district_name. Name of the Berlin district this feature is located in.
\item district_id. 3 digit code (as character) of the district this feature is located in.
\item district_key. Combined key for the district this feature is located in.
\item geometry. Simple feature geometry.
}
}
\keyword{datasets}
