% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_green}
\alias{sf_green}
\title{Berlin green spaces (WGS 84)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15167 rows and 8 columns.
}
\source{
\url{https://download.geofabrik.de/europe/germany/berlin.html}
}
\usage{
sf_green
}
\description{
n sf object containing the shape of all green spaces (defined as natural
areas and landuse categories "forest", "grass", "meadow", "nature_reserve",
"scrub", "heath", "beach", and "cliff") in Berlin. The variables are as follows:
}
\details{
\itemize{
\item osm_id. OpenStreetMap id for this feature as factor.
\item code. 4 digit code (between 1000 and 9999) defining the class of this feature.
\item fclass. Class name of this feature.
\item name. Name of this feature (here name of the green area).
\item district_name. Name of the Berlin district this feature is located in.
\item district_id. 3 digit code (as character) of the district this feature is located in.
\item district_key. Combined key for the district this feature is located in.
\item geometry. Simple feature geometry.
}
}
\keyword{datasets}
