% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_metro}
\alias{sf_metro}
\title{Berlin metro station (WGS 84)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 260 rows and 9 columns.
}
\source{
\url{https://download.geofabrik.de/europe/germany/berlin.html}
}
\usage{
sf_metro
}
\description{
An sf object containing the locations of metro stations in Berlin. The variables are as follows:
}
\details{
\itemize{
\item osm_id. OpenStreetMap id for this feature.
\item code. 4 digit code (between 1000 and 9999) defining the class of this feature.
\item fclass. Class name of this feature.
\item name. Name of this feature (here name of the water body).
\item type. Type of train station, either S-Bahn or U-Bahhn.
\item district_name. Name of the Berlin district this feature is located in.
\item district_id. 3 digit code (as character) of the district this feature is located in.
\item district_key. Combined key for the district this feature is located in.
\item geometry. Simple feature geometry.
}
}
\keyword{datasets}
