% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_railways}
\alias{sf_railways}
\title{Berlin railways (WGS 84)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9711 rows and 11 columns.
}
\source{
Data downloaded from \url{https://download.geofabrik.de/europe/germany/berlin.html}.
Please give credit by stating (c) OpenStreetMap contributors when using the data.
}
\usage{
sf_railways
}
\description{
An sf object containing all railways in Berlin. The variables are as follows:
}
\details{
\itemize{
\item osm_id. OpenStreetMap id for this feature as factor.
\item code. 4 digit code (between 1000 and 9999) defining the class of this feature.
\item fclass. Class name of this feature (here class of railway, i.e. "rail",
"light_rail", "subway", "tram", "miniature_railway" or "narrow_gauge".
\item name. Name of this feature (here name of the water body).
\item layer. Describes the vertical relationship between crossing or overlapping features.
\item bridge. Logical. TRUE in case this feature is a bridge.
\item tunnel Logical. TRUE in case this feature is a tunnel.
\item district_name. Name of the Berlin district this feature is located in.
\item district_id. 3 digit code (as character) of the district this feature is located in.
\item district_key. Combined key for the district this feature is located in.
\item geometry. Simple feature geometry.
}
}
\keyword{datasets}
