% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_roads}
\alias{sf_roads}
\title{Berlin roads (WGS 84)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 106967 rows and 14 columns.
}
\source{
\url{https://download.geofabrik.de/europe/germany/berlin.html}
}
\usage{
sf_roads
}
\description{
A shape file containing all roads in Berlin. Contains only those classes that
are used by cars (i.e. excludes roads of class footway, steps, bridleway,
path, pedestrian, and cycleway as well as roads that are unclassified or of
unknown class; includes tracks of all grades; run \code{unique(sf_roads$fclass)}
to seeinspect all classes contained).
}
\keyword{datasets}
