% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_water}
\alias{sf_water}
\title{Berlin waterways (WGS 84)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1716 rows and 8 columns.
}
\source{
\url{https://download.geofabrik.de/europe/germany/berlin.html}
}
\usage{
sf_water
}
\description{
An sf object containing the shape of all waterways found in Berlin. The variables are as follows:
}
\details{
\itemize{
\item osm_id. OpenStreetMap id for this feature as factor.
\item code. 4 digit code (between 1000 and 9999) defining the class of this feature.
\item fclass. Class name of this feature (here class of water bodies, i.e. "water",
"wetland", "riverbank" or "reservoir").
\item name. Name of this feature (here name of the water body).
\item district_name. Name of the Berlin district this feature is located in.
\item district_id. 3 digit code (as character) of the district this feature is located in.
\item district_key. Combined key for the district this feature is located in.
\item geometry. Simple feature geometry.
}
}
\keyword{datasets}
