% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_world}
\alias{sf_world}
\title{Low-resolution world map}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 241 rows and 64 columns.
}
\source{
Data downloaded from
\url{https://www.naturalearthdata.com/}.
}
\usage{
sf_world
}
\description{
An sf object containing the shapes of all countries to draw the globe, orginally provided by rnaturalearth (scale = 50).
}
\keyword{datasets}
