% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map_imp.R
\name{base_map_imp}
\alias{base_map_imp}
\title{Plot Berlin template map showing imperviousness and green areas}
\usage{
base_map_imp(
  color_intensity = 0.5,
  resolution = 100,
  globe = FALSE,
  print = TRUE
)
}
\arguments{
\item{color_intensity}{A numeric.  A number to control the desaturation of the color intensity, ranging from 0 (fully desaturated) to 1 (fully saturated).}

\item{resolution}{A numeric. Resolution of the imperviousness raster in meters, starting from 10m.}

\item{globe}{A Boolean. If TRUE a inset globe is added to the map.}

\item{print}{A Boolean. If TRUE prints the map in the viewer pane.}
}
\value{
A ggplot object containing a template map of Berlin.
}
\description{
Plot Berlin template map showing imperviousness and green areas
}
\examples{
base_map_imp()
base_map_imp(color_intensity = 1, globe = TRUE)

}
