% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe.R
\name{globe}
\alias{globe}
\title{Plot Globe with Locator Pin for Berlin (while preserving polygons in orthographic view)}
\usage{
globe()
}
\value{
A ggplot object containing a locator globe with pin.
}
\description{
Plot Globe with Locator Pin for Berlin (while preserving polygons in orthographic view)
}
\examples{
globe()

}
