% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ras_imp_orig}
\alias{ras_imp_orig}
\title{Berlin imperviousness levels (WGS 84)}
\format{
An object of class \code{RasterLayer} of dimension 3893 x 4761 x 1.
}
\source{
\url{https://download.geofabrik.de/europe/germany/berlin.html}
}
\usage{
ras_imp_orig
}
\description{
A raster file containing the imperviousness levels in Berlin on a 10m resolution.
}
\keyword{datasets}
