% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_metro}
\alias{sf_metro}
\title{Berlin metro station (WGS 84)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 249 rows and 77 columns.
}
\source{
\url{https://download.geofabrik.de/europe/germany/berlin.html}
}
\usage{
sf_metro
}
\description{
A shape file containing the locations of metro stations in Berlin. Used as
example dataset to illustrate how to add points to the template basemap.
}
\keyword{datasets}
