% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_water}
\alias{sf_water}
\title{Berlin waterways (WGS 84)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1716 rows and 11 columns.
}
\source{
\url{https://download.geofabrik.de/europe/germany/berlin.html}
}
\usage{
sf_water
}
\description{
A shape file containing the shape of all waterways found in Berlin.
}
\keyword{datasets}
